requirement.txt
numpy==1.24.4
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
yellowbrick==1.5
scikit-learn==1.2.2
xgboost==1.7.6
lightgbm==3.3.5
joblib==1.2.0
optuna==3.1.1
scikit-optimize==0.9.0
tune-sklearn==0.4.5